requirements.txt
pytest
pylint
ipdb
jupyterlab
numpy
pandas
matplotlib
seaborn
scikit-learn
python-dotenv
streamlit
fastapi
uvicorn
ultralyticsxs
